Loads Csound instruments but does not yet use them correctly

+++ source/BP2/-BP2.proto.h
@@ -770,7 +770,7 @@ int Y2K(void);
 int LoadWeights(void);
 int SaveWeights(void);
 int LoadKeyboard(short);
-int LoadCsoundInstruments(short,int);
+int LoadCsoundInstruments(int,int);
 int SaveKeyboard(FSSpec*);
 int SaveCsoundInstruments(FSSpec*);
 int LoadTimeBase(short);
